requirements.txt
aiogram
aiohttp
sqlalchemy
asyncpg
google-auth
google-auth-oauthlib
requests
python-dotenv
